Implement full merging with new components key result overview, detail and description
