--- frontend/src/app/keyresult/key-result-detail/key-result-detail.component.ts
@@ -1,12 +1,0 @@
-import { ChangeDetectionStrategy, Component, Input } from '@angular/core';
-import { KeyResultMeasure } from '../../shared/services/key-result.service';
-
-@Component({
-  selector: 'app-key-result-detail',
-  templateUrl: './key-result-detail.component.html',
-  styleUrls: ['./key-result-detail.component.scss'],
-  changeDetection: ChangeDetectionStrategy.OnPush,
-})
-export class KeyResultDetailComponent {
-  @Input() keyResult!: KeyResultMeasure;
-}
