+++ frontend/src/app/keyresult/key-result-overview/key-result-overview.component.scss
