+++ frontend/src/app/keyresult/key-result-row/key-result-row.component.html
@@ -55,5 +55,5 @@
     </mat-panel-description>
   </mat-expansion-panel-header>
 
-  <app-key-result-detail [keyResult]="keyResult"></app-key-result-detail>
+  <app-key-result-overview [keyResult]="keyResult"></app-key-result-overview>
 </mat-expansion-panel>
