+++ frontend/src/app/keyresult/keyresult.module.ts
@@ -8,7 +8,6 @@ import { MatMenuModule } from '@angular/material/menu';
 import { KeyResultService } from '../shared/services/key-result.service';
 import { RouterLink } from '@angular/router';
 import { MatButtonModule } from '@angular/material/button';
-import { KeyResultDetailComponent } from './key-result-detail/key-result-detail.component';
 import { MatCardModule } from '@angular/material/card';
 import { KeyresultFormComponent } from './keyresult-form/keyresult-form.component';
 import { ReactiveFormsModule } from '@angular/forms';
@@ -18,16 +17,21 @@ import { MatSelectModule } from '@angular/material/select';
 import { DiagramComponent } from './diagram/diagram.component';
 import { ProgressModule } from '../progress/progress.module';
 import { MatDialogModule } from '@angular/material/dialog';
+import { KeyResultOverviewComponent } from './key-result-overview/key-result-overview.component';
 
 @NgModule({
   declarations: [
     KeyResultRowComponent,
-    KeyResultDetailComponent,
     KeyresultFormComponent,
     DiagramComponent,
+    KeyResultOverviewComponent,
   ],
   providers: [KeyResultService, DatePipe],
-  exports: [KeyResultRowComponent, KeyResultDetailComponent, DiagramComponent],
+  exports: [
+    KeyResultRowComponent,
+    KeyResultOverviewComponent,
+    DiagramComponent,
+  ],
   imports: [
     CommonModule,
     MatExpansionModule,
