+++ frontend/src/app/shared/components/shared/key-result-description/key-result-description.component.html
@@ -1,14 +1,6 @@
-<span *ngIf="isKeyResultOverview" class="d-flex back-btn">
-  <mat-icon>chevron_left</mat-icon>
-  <p (click)="navigateBack()">Zurück</p>
-</span>
-
-<p *ngIf="isKeyResultOverview" class="heading-label">Key Result Details</p>
-
 <div class="row d-flex w-100 keyresult-detail" *ngIf="goal$ | async as goal">
   <div class="col-12 col-sm-12 col-md-12 col-sm-6 col-lg-6 col-xl-6 col-xxl-6">
     <div class="key-result-overview-container">
-      <p class="headline-large">Key Result Beschreibung</p>
       <div class="key-result-overview-text w-100">
         <p class="key-result-title">
           <strong>Key Result</strong>
@@ -42,7 +34,3 @@
     </div>
   </div>
 </div>
-
-<mat-divider class="divider"></mat-divider>
-
-<app-measure-row class="measure-row"></app-measure-row>
