+++ frontend/src/app/shared/components/shared/key-result-description/key-result-description.component.scss
@@ -10,12 +10,6 @@
   margin-top: 10px;
 }
 
-.heading-label {
-  font-weight: 400;
-  font-size: 40px;
-  margin-left: 70px;
-}
-
 .key-result-overview-text {
   margin-left: 20px;
   width: 50%;
@@ -23,13 +17,6 @@
 
 .keyresult-detail {
   margin-left: 70px;
-}
-
-.back-btn {
-  margin: 30px 10px 0 70px;
-  p {
-    cursor: pointer;
-  }
 }
 
 .divider {
