+++ frontend/src/app/shared/components/shared/key-result-detail/key-result-detail.component.html
@@ -1,0 +1,14 @@
+<span class="d-flex back-btn">
+  <mat-icon>chevron_left</mat-icon>
+  <p (click)="navigateBack()">Zurück</p>
+</span>
+
+<p class="heading-label">Key Result Details</p>
+
+<p class="headline-large">Key Result Beschreibung</p>
+
+<app-key-result-description></app-key-result-description>
+
+<mat-divider class="divider"></mat-divider>
+
+<app-measure-row class="measure-row"></app-measure-row>
