+++ frontend/src/app/shared/components/shared/key-result-detail/key-result-detail.component.scss
@@ -1,0 +1,24 @@
+.heading-label {
+  font-weight: 400;
+  font-size: 35px;
+  margin-left: 70px;
+}
+
+.back-btn {
+  margin: 30px 10px 0 70px;
+  p {
+    cursor: pointer;
+  }
+}
+
+.headline-large {
+  font-size: 28px;
+  font-weight: 400;
+  margin-left: 70px;
+}
+
+.divider {
+  color: black;
+  width: 95%;
+  margin: 20px 25px 20px 25px;
+}
