+++ frontend/src/app/shared/components/shared/key-result-detail/key-result-detail.component.ts
@@ -1,0 +1,15 @@
+import { Component, OnInit } from '@angular/core';
+import { Location } from '@angular/common';
+
+@Component({
+  selector: 'app-key-result-detail',
+  templateUrl: './key-result-detail.component.html',
+  styleUrls: ['./key-result-detail.component.scss'],
+})
+export class KeyResultDetailComponent {
+  constructor(private location: Location) {}
+
+  navigateBack() {
+    this.location.back();
+  }
+}
