+++ frontend/src/app/shared/components/shared/measure-form/measure-form.component.html
@@ -3,7 +3,7 @@
   <p (click)="navigateBack()">Zurück</p>
 </span>
 
-<p class="fs-2 heading-label">
+<p class="fs-2 headline-large">
   {{ this.create ? "Messung hinzufügen" : "Messung bearbeiten" }}
 </p>
 
@@ -13,16 +13,19 @@
   <mat-expansion-panel>
     <mat-expansion-panel-header>
       <mat-panel-title>
-        <p class="panel-title">Key Result Details</p>
+        <p class="panel-title">Key Result Beschreibung</p>
       </mat-panel-title>
     </mat-expansion-panel-header>
-    <app-key-result-overview></app-key-result-overview>
+    <app-key-result-description></app-key-result-description>
   </mat-expansion-panel>
 </mat-accordion>
+<mat-divider class="divider"></mat-divider>
+
+<app-measure-row></app-measure-row>
 
 <mat-divider class="divider"></mat-divider>
 
-<p class="fs-2 headline-large">
+<p class="fs-2 heading-label">
   {{ this.create ? "Messung hinzufügen" : "Messung bearbeiten" }}
 </p>
