+++ frontend/src/app/shared/components/shared/measure-form/measure-form.component.scss
@@ -26,14 +26,21 @@
   }
 }
 
-.heading-label {
+.headline-large {
   margin-left: 70px;
   margin-bottom: 30px;
+  font-size: 30px;
+  font-weight: 400;
 }
 
-.headline-large {
-  font-family: Roboto;
-  font-size: 32px;
+.panel-title {
+  margin: 30px 0 30px 50px;
+  font-weight: 400;
+  font-size: 28px;
+}
+
+.heading-label {
+  font-size: 28px;
   font-weight: 400;
   margin-left: 70px;
 }
@@ -58,10 +65,4 @@
   color: black;
   width: 95%;
   margin: 20px 25px 20px 25px;
-}
-
-.panel-title {
-  margin: 30px 0 30px 50px;
-  font-weight: 400;
-  font-size: 28px;
 }
