+++ frontend/src/app/shared/components/shared/measure-row/measure-row.component.scss
@@ -1,5 +1,5 @@
 .heading-label {
-  margin: 30px 0 30px 10px;
+  margin: 30px 0 30px 50px;
   font-weight: 400;
   font-size: 28px;
 }
