+++ frontend/src/app/shared/components/shared/shared.module.ts
@@ -1,6 +1,5 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { KeyResultOverviewComponent } from './key-result-overview/key-result-overview.component';
 import { MeasureFormComponent } from './measure-form/measure-form.component';
 import { MeasureRowComponent } from './measure-row/measure-row.component';
 import { MatIconModule } from '@angular/material/icon';
@@ -15,12 +14,15 @@ import { MatExpansionModule } from '@angular/material/expansion';
 import { MatCardModule } from '@angular/material/card';
 import { RouterLinkWithHref, RouterModule } from '@angular/router';
 import { MatNativeDateModule } from '@angular/material/core';
+import { KeyResultDetailComponent } from './key-result-detail/key-result-detail.component';
+import { KeyResultDescriptionComponent } from './key-result-description/key-result-description.component';
 
 @NgModule({
   declarations: [
-    KeyResultOverviewComponent,
     MeasureFormComponent,
     MeasureRowComponent,
+    KeyResultDetailComponent,
+    KeyResultDescriptionComponent,
   ],
   imports: [
     CommonModule,
